Move example files constant to domain layer (where it belongs)
