+++ timetable_solutions/tests/unit/interfaces/data_upload/test_example_file_downloads.py
@@ -7,7 +7,7 @@ from django import test
 from django import urls
 
 # Local application imports
-from constants.example_files import ExampleFile
+from domain.data_upload_processing.constants import ExampleFile
 from constants.url_names import UrlName
